requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
Django==3.0.6
django-js-asset==1.2.2
feedfinder2==0.0.4
feedparser==5.2.1
html5lib==1.0.1
idna==2.9
jieba3k==0.35.1
joblib==0.15.1
lxml==4.5.1
newspaper3k==0.2.8
nltk==3.5
Pillow==7.1.2
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.7
reportlab==3.5.42
requests==2.23.0
requests-file==1.5.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.1
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
